JWTfunctionAndDeployRdy

+++ src/main/java/com/AskABot/AskABot/controller/TopicController.java
@@ -1,6 +1,8 @@
 package com.AskABot.AskABot.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,6 +14,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.AskABot.AskABot.model.Topic;
 import com.AskABot.AskABot.service.TopicService;
+
+import jakarta.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping("/topic")
@@ -26,9 +30,13 @@ public class TopicController {
     }
 
     @PostMapping("/createTopic")
-    public Topic createTopic(@RequestBody Topic topic) {
+    public Topic createTopic(@RequestBody Topic topic, HttpServletRequest request) {
+        String userId = (String) request.getAttribute("userId");
+        if (userId == null) {
+            throw new IllegalArgumentException("User ID is missing");
+        }
+        topic.setCreatedByUser(userId);
         return topicService.createTopic(topic);
-        
     }
 
     @GetMapping("/getTopics")
@@ -37,7 +45,11 @@ public class TopicController {
     }
 
     @DeleteMapping("/deleteTopic")
-    public void deleteTopic(@RequestParam String topicId, @RequestParam String userId) {
+    public void deleteTopic(@RequestParam String topicId, @RequestParam String userId2, HttpServletRequest request) {
+        String userId = (String) request.getAttribute("userId");
+        if (userId == null) {
+            throw new IllegalArgumentException("User ID is missing");
+        }
         topicService.deleteTopic(topicId, userId);
     }
